+++ src/net/sourceforge/sheltermanager/asm/bo/AnimalVaccination.java
@@ -143,11 +143,11 @@ public class AnimalVaccination extends UserInfoBO<AnimalVaccination> {
         rs.setField("DateRequired", newValue);
     }
 
-    public Double getCost() throws CursorEngineException {
-        return new Double(rs.getDouble("Cost"));
+    public Integer getCost() throws CursorEngineException {
+        return new Integer(rs.getInt("Cost"));
     }
 
-    public void setCost(Double newValue) throws CursorEngineException {
+    public void setCost(Integer newValue) throws CursorEngineException {
         rs.setField("Cost", newValue);
     }
