+++ src/net/sourceforge/sheltermanager/asm/bo/MedicalProfile.java
@@ -80,11 +80,11 @@ public class MedicalProfile extends UserInfoBO<MedicalProfile> {
         rs.setField("TimingRule", newValue);
     }
 
-    public Double getCost() throws CursorEngineException {
-        return new Double(rs.getDouble("Cost"));
+    public Integer getCost() throws CursorEngineException {
+        return new Integer(rs.getInt("Cost"));
     }
 
-    public void setCost(Double newValue) throws CursorEngineException {
+    public void setCost(Integer newValue) throws CursorEngineException {
         rs.setField("Cost", newValue);
     }
